+++ src/data_model/utils.jl
@@ -653,7 +653,7 @@ function _add_gen_cost_model!(math_obj::Dict{String,<:Any}, eng_obj::Dict{String
     math_obj["model"] = get(eng_obj, "cost_pg_model", 2)
     math_obj["startup"] = 0.0
     math_obj["shutdown"] = 0.0
-    math_obj["cost"] = get(eng_obj, "cost_pg_parameters", [0.0, 1.0, 0.0])
+    math_obj["cost"] = get(eng_obj, "cost_pg_parameters", [1.0, 0.0])
     math_obj["ncost"] = length(math_obj["cost"])
 end